functions/requirements.txt
Flask
flask-cors
firebase-admin
requests
Pillow
numpy
gunicorn